modified CRC and Roboclaw classes for testing

+++ java/CRC16.java
@@ -28,15 +28,15 @@ public class CRC16 {
    * @param packet the data for us to calculate the crc to
    * @return crc long value;
    */
-  public static long calcCrc16(byte[] packet) {
-    long crc = 0x0000;
-    long topBit = 1L << 15;
-    long mask = (topBit << 1) - 1;
+  public static int calcCrc16(byte[] packet) {
+    int crc = 0x0000;
+    int topBit = 1L << 15;
+    int mask = (topBit << 1) - 1;
 
     for (int i = 0; i < packet.length; i++) {
-      long curByte = reflect((long) packet[i] & 0x00FFL, 8);
+      int curByte = reflect((long) packet[i] & 0x00FFL, 8);
       for (int j = 0x80; j != 0; j >>= 1) {
-        long bit = crc & topBit;
+        int bit = crc & topBit;
         crc <<= 1;
         if ((curByte & j) != 0) {
           bit ^= topBit;
